Sales Database of a Superstore (Data Querry)
A Superstore is a very large supermarket or retail shop that sells various categories of
household goods, clothes, electrical appliances and devices, office items and 
equipment and accessories as well as grosery. A wide variety of goods and
merchandise displayed for sale are often sold at a discount. The supermamrket serves
as a one-stop destination for shoppers.
For the purpose of this project, our superstore database consist of the sales and
shipment of some categories of items, with goods grouped as sub-categories under
them.
The category include, furniture, office supplies and technologies further sub, consist of
Bookcases, chairs, labels, tables, storage, furnishings, Art, phones, binders,
appliances, paper, binders, accessories, labels, fasteners, envelops, copiers and
supplies.
The Superstore database consist of four tables each with different columns. The Columns include CUSTOMER, ORDER, PRODUCT and REGION.